added game configs and game tracking system for valorant overwatch counter strike 2 , updated websockets and added google oauth in backend.

+++ backend/src/main/java/com/Questboard/backend/infrastructure/GameEventRequest.java
@@ -7,16 +7,6 @@ import java.util.UUID;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
-// @Data
-// public class GameEventRequest {
-//     private UUID userId;
-//     private String gameSlug;
-//     private Integer gameId;       // Overwolf Game ID (e.g., 21566 for Valorant)
-//     private String eventType;     // e.g., "KILL", "MATCH_WIN"
-//     private String eventTimestamp; // To prevent replay attacks
-//     private Map<String, Object> rawData; // Entire JSON block from Overwolf GEP
-// }
-
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class GameEventRequest {
@@ -25,13 +15,13 @@ public class GameEventRequest {
 
     private String gameSlug;
 
-    private Integer gameId;
+    private Long gameId;
 
     private String eventType;
 
     private String eventTimestamp;
 
-    private Integer count;
+    private Long count;
 
     private Map<String, Object> metadata;
